improvement(monitors): refactor into a feature partition
