+++ packages/shared/src/features/monitors/index.ts
@@ -1,0 +1,7 @@
+/** index.ts is the client-safe barrel for the monitors feature — domain
+ * schemas, types, refinements, and standalone predicates. Server-only code
+ * (the service, scheduler DTOs) is re-exported from `./server` instead. */
+export * from "./types";
+export * from "./isValidQuery";
+export * from "./isValidTemplate";
+export * from "./isValidThresholdOrder";
