+++ packages/shared/src/features/monitors/isValidTemplate.test.ts
