+++ packages/shared/src/features/monitors/server.ts
@@ -1,0 +1,6 @@
+/** server.ts is the server-only barrel for the monitors feature. It
+ * re-exports everything client-safe from `./index` plus the service and
+ * scheduler DTOs. */
+export * from "./index";
+export * from "./service";
+export * from "./scheduler/types";
