+++ packages/shared/src/features/monitors/service/index.ts
@@ -1,0 +1,5 @@
+/** service/index.ts is the MonitorService barrel — public DTOs and the
+ * service class. Mapping/calculate helpers in `./helpers` are intentionally
+ * not re-exported. */
+export * from "./service";
+export * from "./types";
