+++ packages/shared/src/server/index.ts
@@ -112,7 +112,7 @@ export * from "./redis/llmAsJudgeExecutionQueue";
 export * from "./services/sessions-ui-table-service";
 export * from "./services/sessions-ui-table-events-service";
 export * from "./services/DashboardService";
-export * from "./services/MonitorService";
+export * from "../features/monitors/server";
 export * from "./services/TableViewService";
 export * from "./services/DefaultViewService";
 export * from "./services/DefaultEvaluationModelService";
