--- packages/shared/src/server/services/MonitorService/index.ts
@@ -1,4 +1,0 @@
-export * from "./MonitorService";
-export * from "./isValidTemplate";
-export * from "./types";
-export * from "./isValidQuery";
